Revert "chore: moved type files to type folder"
This reverts commit f5d7a95e5a0221c4670061c45ef0f5f27ca46599.
