+++ packages/data-layer/src/index.ts
@@ -1,5 +1,7 @@
 export { DataLayer } from "./data-layer";
 export { useDataLayer, DataLayerContext, DataLayerProvider } from "./react";
 export * from "./openapi-search-client/models/index";
-export * from "./types";
+export * from "./data.types";
+export * from "./roundApplication.types";
+export * from "./attestations.types";
 export * from "./utils";
